requirements.txt
snowflake-connector-python==3.6.0
python-dateutil==2.8.2